Update Antlr4 and perform maintenance (fix warnings)
LogicScript has seemingly not been maintained for a while. This commit will perform some maintenance, so that a project using it has better support.

Originally this commit was only meant to fix/suppress warnings. In that process however I noticed that LogicScript was already suppressing a warning for incompatible netstandard versions - which you get when using a new net core version like net8.0 and newer. And that was cause by a very old antlr4 version.

This commit will:
- Update Antlr4 to run the latest 'Antlr4.Runtime.Standard' version. This compiles for 'netstandard2.0' and will also fix Antlr4 integrity issues in JetBrains Rider IDE. As consequence I had to update the 'SyntaxError' handler method - nothing else.
- Fix all compiler warnings. I had to remove a 'MemberNotNull' attribute and a nullable indicator - as they just did not play nice. And added a bunch of auto-generated equality code.
- Removed unused using statements. Cause I was already cleaning up.